XWIKI-11205: If view rights are not allowed for guest users then ResetPassword doesn't work
Fixes some problems after first review of code:
  * don't use nested class anymore
  * add a programming right check
  * small optimization to inject a component with a provider
  * small other fixes related to style
