+++ xwiki-platform-core/xwiki-platform-flamingo/xwiki-platform-flamingo-skin/xwiki-platform-flamingo-skin-resources/src/main/resources/flamingo/layoutvars.vm
@@ -37,8 +37,9 @@
   #set($showRightPanels = $xwiki.getXWikiPreference("showRightPanels"))
   #set($leftPanelsWidth = $xwiki.getXWikiPreference("leftPanelsWidth"))
   #set($rightPanelsWidth = $xwiki.getXWikiPreference("rightPanelsWidth"))
-#elseif ("$!doc" != "") ## We only rely on the space reference if we are displaying a doc
-                        ## we want to be able to control this value in case of special resources such as authenticate.
+## We only rely on the space reference if we are displaying a doc
+## we want to be able to control this value in case of special resources such as authenticate.
+#elseif ("$!doc" != "")
   #set($showLeftPanels = $xwiki.getSpacePreference("showLeftPanels"))
   #set($showRightPanels = $xwiki.getSpacePreference("showRightPanels"))
   #set($leftPanelsWidth = $xwiki.getSpacePreference("leftPanelsWidth"))
