+++ xwiki-platform-core/xwiki-platform-security/xwiki-platform-security-authentication/xwiki-platform-security-authentication-api/src/main/java/org/xwiki/security/authentication/AuthenticationResourceReference.java
@@ -17,9 +17,8 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.xwiki.security.authentication.api;
+package org.xwiki.security.authentication;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.xwiki.resource.AbstractResourceReference;
@@ -45,57 +44,6 @@ public class AuthenticationResourceReference extends AbstractResourceReference
      * Dedicated resource type used for authentication.
      */
     public static final ResourceType TYPE = new ResourceType(RESOURCE_TYPE_ID);
-
-    /**
-     * Possible actions for authentication resource reference.
-     */
-    public enum AuthenticationAction
-    {
-        /**
-         * Action used to reset the password of a user.
-         */
-        RESET_PASSWORD("reset"),
-
-        /**
-         * Action used to retrieve the username of a user.
-         */
-        FORGOT_USERNAME("forgot");
-
-        private String requestParameter;
-
-        AuthenticationAction(String requestParameter)
-        {
-            this.requestParameter = requestParameter;
-        }
-
-        /**
-         * @return the request parameter associated to this action.
-         */
-        public String getRequestParameter()
-        {
-            return requestParameter;
-        }
-
-        /**
-         * Retrieve an action based on the request parameter.
-         * @param parameter the parameter from which to retrieve the action.
-         * @return an action associated to the given parameter.
-         * @throws IllegalArgumentException if the parameter is empty or if no action could be found based on it.
-         */
-        public static AuthenticationAction getFromRequestParameter(String parameter) throws IllegalArgumentException
-        {
-            if (StringUtils.isEmpty(parameter)) {
-                throw new IllegalArgumentException("The parameter needs to be provided.");
-            }
-            for (AuthenticationAction value : AuthenticationAction.values()) {
-                if (value.requestParameter.equals(parameter.toLowerCase())) {
-                    return value;
-                }
-            }
-            throw new IllegalArgumentException(
-                String.format("Cannot find an AuthenticationAction for parameter [%s]", parameter));
-        }
-    }
 
     private AuthenticationAction action;
